Add trigger for cloudfront cache bust
This adds an option to allow users to trigger a CloudFront
invalidation. We have to pass in the ID of the
distribution to bust, which to me isn't ideal, but hey it's
better than having to log into the console when the whole
point of having quasimodo is so that you don't have to do that.